requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.2.0
rank-bm25==0.2.2
nltk==3.8.1
scikit-learn==1.3.1
python-multipart==0.0.6
twilio==8.10.0
httpx==0.25.0
pytest==7.4.2
